requirements.txt

pandas==2.2.3
numpy==2.2.4
SQLAlchemy==2.0.40
python-dotenv==1.1.0
openai==1.72.0
langchain==0.3.23
langchain-openai==0.3.12
pydantic==2.11.3
tqdm==4.67.1